+++ src/Managers/Enums/VersionResolveStrategy.cs
@@ -1,0 +1,8 @@
+﻿namespace SeleniumSharper.Managers.Enums;
+
+public enum VersionResolveStrategy
+{
+    LatestVersion,
+    InstalledVersion,
+    SpecificVersion
+}